requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
Werkzeug==3.1.3
